test(Root.spec.js): revert expect -> chai.expect; all Root tests pass
